requirements.txt
pygame==2.1.3
pyserial==3.5
